requirements.txt
PyGithub==1.54.1
GitPython==3.1.14